+++ src/MicroService.ApiGateway.Domain/Entites/Ocelot/Headers.cs
@@ -4,7 +4,7 @@ namespace MicroService.ApiGateway.Entites.Ocelot
 {
     public class Headers : Entity<int>
     {
-        public virtual int ReRouteId { get; private set; }
+        public virtual long ReRouteId { get; private set; }
         public virtual string Key { get; private set; }
         public virtual string Value { get; private set; }
 
@@ -12,7 +12,7 @@ namespace MicroService.ApiGateway.Entites.Ocelot
         {
 
         }
-        public Headers(int rerouteId)
+        public Headers(long rerouteId)
         {
             ReRouteId = rerouteId;
         }
